PPO-ASV-path-follow/src/requirements_tf_291.txt
tf-agents==0.13.0
scipy==1.9.0
gym
matplotlib==3.5.3
numpy==1.22.4
tensorflow-probability==0.17.0
